cleanup and styling
