+++ frontend/src/Components/PlenumsCardLight.tsx
@@ -34,7 +34,7 @@ export default function PlenumsCardLight(props:Readonly<PlenumsCardLightProps>) 
     }
 
     return (
-        <div id="card" className={"plenumsCardLight w-50 p-2 mb-5"} key={props.plenum.id} style={cardStyle}>
+        <div id="card" className={"plenumsCardLight w-55 p-4 mb-5"} key={props.plenum.id} style={cardStyle}>
             <div className={"cardGroup text-2xl"}>{props.plenum.group}</div>
             <div className={"text-2xl"}>{(new Date(Date.parse(props.plenum.date))).toLocaleDateString("de-EU")}</div>
         </div>
