+++ frontend/src/Pages/Events.tsx
@@ -5,7 +5,7 @@ export default function Events(){
     return(
         <>
             <header className={"topBar"}><Navbar/></header>
-            <h1>Events</h1>
+            <h1 className={"text-2xl font-semibold"}>Coming soon..</h1>
         </>
     )
 }
