+++ frontend/src/Pages/PlenumsDetails.tsx
@@ -39,13 +39,13 @@ export default function PlenumsDetails() {
         <>
             <header className={"topBar"}><Navbar/></header>
         <div className={"mainContainer"}>
-           <div className={"rightContainer"}>
+           <div className={""}>
                <div>
                    {plenum ? <PlenumsCard plenum={plenum} updateToggle={updateToggle}/> : <div>loading..</div>}
                </div>
             </div>
-            <div id="updateDiv" className="bor">
-                <div  className={"mainContainer"}>
+            <div id="updateDiv" className="">
+                <div  className={""}>
                     {plenum? <PlenumUpdateForm plenum={plenum} onUpdate={getPlenumById}/> : <div>loading..</div>}
                 </div>
             </div>
