+++ frontend/src/Pages/Werki.tsx
@@ -5,7 +5,7 @@ export default function Werki(){
     return(
         <>
             <header className={"topBar"}><Navbar/></header>
-            <h1>Werki</h1>
+            <h1 className={"text-2xl font-semibold"}>Coming soon..</h1>
         </>
     )
 }
