+++ frontend/src/Styles/index.css
@@ -13,7 +13,7 @@ h1{
 }
 
 button {
-    background-color: darkred;
+    background-color: #af3a3a;
     color: white;
     padding: 10px;
     margin: 10px;
